refactor: move middleware to routes directory
