+++ src/routes/middleware/RequirePatron.ts
@@ -1,9 +1,9 @@
 import express, { NextFunction } from 'express';
 
-import TokenRepository from '../data_layer/TokenRepository';
-import UsersRepository from '../data_layer/UsersRepository';
-import AuthenticationService from '../services/AuthenticationService';
-import { getDatabase } from '../data_layer';
+import TokenRepository from '../../data_layer/TokenRepository';
+import UsersRepository from '../../data_layer/UsersRepository';
+import AuthenticationService from '../../services/AuthenticationService';
+import { getDatabase } from '../../data_layer';
 
 const RequirePatron = async (
   req: express.Request,
